# docs/00-product-baseline/contracts/openapi-v2-p0.yaml
openapi: 3.1.0
info:
  title: Ikaros V2 P0 API
  version: 0.1.0
  description: |
    Machine-readable P0 HTTP contract for Ikaros V2.
    Business semantics are owned by docs/v2/contracts/P0-Command-Query-Event-Catalog.md.
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /media/subjects:
    get:
      tags: [Media]
      operationId: listMediaSubjects
      x-ikaros-contract-id: media.list-subjects
      summary: List owned video subjects
      parameters:
        - name: X-Ikaros-Actor-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Video subjects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MediaSubject' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Media]
      operationId: createMediaSubject
      x-ikaros-contract-id: media.create-subject
      summary: Create a video subject and its Resource
      parameters:
        - name: X-Ikaros-Actor-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, kind]
              properties:
                title: { type: string, minLength: 1, maxLength: 512 }
                kind: { type: string, enum: [SERIES, MOVIE, VIDEO] }
                locale: { type: string, minLength: 1, maxLength: 32, default: en-US }
      responses:
        '200':
          description: Created video subject
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MediaSubject' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /media/subjects/{subject_id}/seasons/{season_id}/actions/reorder-episodes:
    post:
      tags: [Media]
      operationId: reorderMediaEpisodes
      x-ikaros-contract-id: media.reorder-episodes
      summary: Reorder all episodes in a season
      parameters:
        - name: subject_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: season_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: X-Ikaros-Actor-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [episode_ids]
              properties:
                episode_ids:
                  type: array
                  minItems: 1
                  uniqueItems: true
                  items: { type: string, format: uuid }
      responses:
        '200': { description: Episode order saved }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /admin/roles:
    get:
      tags: [Admin Identity]
      operationId: listRoles
      x-ikaros-contract-id: identity.list-roles
      summary: List roles
      responses:
        '200':
          description: Roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Admin Identity]
      operationId: createRole
      x-ikaros-contract-id: identity.create-role
      summary: Create a platform role
      parameters:
        - name: X-Ikaros-Actor-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRoleRequest' }
      responses:
        '201':
          description: Role created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Role' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/roles/{role_id}/permissions:
    put:
      tags: [Admin Identity]
      operationId: replaceRolePermissions
      x-ikaros-contract-id: identity.replace-role-permissions
      summary: Replace a role's platform permissions
      parameters:
        - name: role_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: X-Ikaros-Actor-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplaceRolePermissionsRequest' }
      responses:
        '200':
          description: Updated role
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Role' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/roles/users/{user_id}/roles/{role_id}:
    post:
      tags: [Admin Identity]
      operationId: assignUserRole
      x-ikaros-contract-id: identity.assign-user-role
      summary: Assign a role to a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: role_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: X-Ikaros-Actor-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Role assigned or already assigned }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/roles/users/{user_id}/roles/{role_id}/actions/revoke:
    post:
      tags: [Admin Identity]
      operationId: revokeUserRole
      x-ikaros-contract-id: identity.revoke-user-role
      summary: Revoke a role from a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: role_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: X-Ikaros-Actor-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Role revoked or already absent }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /music/tracks/{track_id}/associations:
    post:
      tags: [Music]
      operationId: createMusicTrackAssociation
      x-ikaros-contract-id: music.create-track-association
      summary: Associate a recognized music candidate with an artist and album
      parameters:
        - { name: track_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: X-Ikaros-Actor-Id, in: header, required: true, schema: { type: string, format: uuid } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateMusicTrackAssociationRequest' } } } }
      responses:
        '200': { description: Track association, content: { application/json: { schema: { $ref: '#/components/schemas/MusicTrackAssociation' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [Music]
      operationId: listMusicTrackAssociations
      x-ikaros-contract-id: music.list-track-associations
      summary: List artist and album associations for a music track
      parameters:
        - { name: track_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: X-Ikaros-Actor-Id, in: header, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Track associations, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/MusicTrackAssociation' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-or-jwt
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authorization denied
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found or not visible
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Business or idempotency conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        request_id: { type: string }
        trace_id: { type: string }
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
...
    MediaSubject:
      type: object
      required: [id, resource_id, kind, created_at]
      properties:
        id: { type: string, format: uuid }
        resource_id: { type: string, format: uuid }
        kind: { type: string, enum: [SERIES, MOVIE, VIDEO] }
        created_at: { type: string, format: date-time }
...
    CreateMusicTrackAssociationRequest:
      type: object
      required: [candidate_id]
      additionalProperties: false
      properties:
        candidate_id: { type: string, format: uuid }
    MusicTrackAssociation:
      type: object
      required: [id, track_id, candidate_id, artist_id, album_id, edition_id, membership_id, created_at]
      properties:
        id: { type: string, format: uuid }
        track_id: { type: string, format: uuid }
        candidate_id: { type: string, format: uuid }
        artist_id: { type: string, format: uuid }
        album_id: { type: string, format: uuid }
        edition_id: { type: string, format: uuid }
        membership_id: { type: string, format: uuid }
        artist: { type: [string, 'null'] }
        album: { type: [string, 'null'] }
        edition_name: { type: string }
        track_number: { type: [integer, 'null'], minimum: 1 }
        created_at: { type: string, format: date-time }
...
    Role:
      type: object
      required: [id, code, name, built_in, permissions]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        description: { type: [string, 'null'] }
        built_in: { type: boolean }
        permissions:
          type: array
          uniqueItems: true
          items: { type: string }
    CreateRoleRequest:
      type: object
      required: [code, name]
      additionalProperties: false
      properties:
        code: { type: string, pattern: '^[A-Z][A-Z0-9_]*$', maxLength: 96 }
        name: { type: string, minLength: 1, maxLength: 128 }
        description: { type: [string, 'null'], maxLength: 2000 }
    ReplaceRolePermissionsRequest:
      type: object
      required: [permissions]
      additionalProperties: false
      properties:
        permissions:
          type: array
          uniqueItems: true
          items: { type: string }
...