# music/src/main/java/run/ikaros/music/MusicTrackAssociationController.java
package run.ikaros.music;
import jakarta.validation.Valid; import java.util.UUID; import org.springframework.web.bind.annotation.GetMapping; import org.springframework.web.bind.annotation.PathVariable; import org.springframework.web.bind.annotation.PostMapping; import org.springframework.web.bind.annotation.RequestBody; import org.springframework.web.bind.annotation.RequestHeader; import org.springframework.web.bind.annotation.RequestMapping; import org.springframework.web.bind.annotation.RestController; import reactor.core.publisher.Flux; import reactor.core.publisher.Mono;
@RestController @RequestMapping("/api/music/tracks/{trackId}/associations") public class MusicTrackAssociationController { private final MusicTrackAssociationService service; public MusicTrackAssociationController(MusicTrackAssociationService service){this.service=service;} @PostMapping public Mono<MusicTrackAssociationView> create(@RequestHeader("X-Ikaros-Actor-Id") UUID ownerId,@PathVariable UUID trackId,@Valid @RequestBody CreateMusicTrackAssociationRequest request){return service.create(ownerId,trackId,request);} @GetMapping public Flux<MusicTrackAssociationView> list(@RequestHeader("X-Ikaros-Actor-Id") UUID ownerId,@PathVariable UUID trackId){return service.list(ownerId,trackId);} }